requirements.txt
streamlit>=1.28.0
bcrypt>=4.0.0
pyotp>=2.9.0
qrcode>=7.4.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pandas>=2.0.0
matplotlib>=3.8.0
openai>=1.0.0 
plotly==5.16.1 
psycopg2-binary
python-jose # if using GPT-powered insights